feat: add coverage, gas reporter, linting and prettier (#12)
* Add code coverage, gas reporter, linting and prettier

* Lint

* Fix comments max-line-lenght

* Fix eslint config

* Add linting to ci

+++ src/contracts/Migrations.sol
@@ -3,23 +3,23 @@
 pragma solidity 0.8.11;
 
 contract Migrations {
-  address public immutable owner;
-  uint public last_completed_migration;
+    address public immutable owner;
+    uint256 public last_completed_migration;
 
-  constructor() {
-    owner = msg.sender;
-  }
+    constructor() {
+        owner = msg.sender;
+    }
 
-  modifier restricted() {
-    if (msg.sender == owner) _;
-  }
+    modifier restricted() {
+        if (msg.sender == owner) _;
+    }
 
-  function setCompleted(uint completed) public restricted {
-    last_completed_migration = completed;
-  }
+    function setCompleted(uint256 completed) public restricted {
+        last_completed_migration = completed;
+    }
 
-  function upgrade(address new_address) public restricted {
-    Migrations upgraded = Migrations(new_address);
-    upgraded.setCompleted(last_completed_migration);
-  }
+    function upgrade(address new_address) public restricted {
+        Migrations upgraded = Migrations(new_address);
+        upgraded.setCompleted(last_completed_migration);
+    }
 }
